+++ lua/obsidian/commands/new.lua
@@ -3,7 +3,7 @@ return function(client, data)
   ---@type obsidian.Note
   local note
   if data.args:len() > 0 then
-    note = client:new_note(data.args)
+    note = client:create_note { title = data.args }
   else
     local title = vim.fn.input {
       prompt = "Enter title (optional): ",
@@ -11,7 +11,7 @@ return function(client, data)
     if string.len(title) == 0 then
       title = nil
     end
-    note = client:new_note(title)
+    note = client:create_note { title = title }
   end
   client:open_note(note)
 end